Add conditional ops and jmps for for loop
Adds equality conditionals (>=,>,==,<,<=) and two types of JMP op,
1) conditional JUMP_IF_FALSE and 2) unconditional JMP. Both JMP
variants take an absolute address to jump to.

This is enought to implement for (;;) loops!

Also cleaned up debugging with a DEBUG_LOG() macro and fixed how
arguments are passed to native function (previously we were passing
the bottom of the stack... whoops).

In release builds the vm.a comes in at 2848b after this change.